improve profermance views
